mobile/ebitenmobileview: handle the touch actions in input_android.go
The cancellation handling does not need a separate file: put it back into
input_android.go and name the MotionEvent actions with the original names
prefixed with an underscore.
